+++ pages/alber.eibd/templatechecker.customwidgets.php
@@ -1,0 +1,10 @@
+<?php
+$customWidgets = array(
+	'shutter.one' => array(),
+	'shutter.two' => array(),
+	'shutter.three' => array(),
+	'smallshutter.smallshut' => array(),
+	'smallrtr.one' => array(),
+	'weatherjustimage.justimage' => array(),
+);
+?>