+++ pages/fleischer/templatechecker.customwidgets.php
@@ -1,0 +1,8 @@
+<?php
+$customWidgets = array(
+	'device.rtrmacflei' => array(),
+	'device.rtrstell' => array(),
+	'shutter.one' => array(),
+	'multi.music' => array(),
+);
+?>